backend/requirements.txt
fastapi
uvicorn
groq
deepgram-sdk
python-dotenv
websockets
jinja2
python-multipart
elevenlabs
SpeechRecognition
pyaudio
pygame
pyttsx3
